refactor(ui): adjust spacing and layout across multiple components
- Reduce vertical padding in BookingSection and WhyChooseSection
- Simplify FeaturedTourLeaders card layout and update profile link
- Reorder ArticleCard placement in Index and WhyUs pages
- Remove stepper component from PackageDetails itinerary
