+++ src/components/WhyChooseSection.vue
@@ -1,5 +1,5 @@
 <template>
-  <section id="why-us" class="py-20 bg-light-blue text-foreground">
+  <section id="why-us" class="py-10 bg-light-blue text-foreground">
     <div class="max-w-6xl mx-auto px-4 sm:px-6 lg:px-8">
       <div class="text-center mb-12">
         <h2 class="text-4xl font-bold text-foreground mb-4">Why Us?</h2>
